cx_patches/src/main/java/org/kuali/kra/bo/Rolodex.java.patch
Commit: edf06332fe7d3cb3940e9353434523fab5d25033
Date:   2014-06-30 11:43:17 -0700 (5 months ago)

Patching fix from changeset 19864. This patch adds back the isSponsorAddress property to Rolodex. When isSponsorAddress is
"N", the behavior changes so the Rolodex reflects the Organization address. For this, organizations had to be added to
Rolodex. The maintainable also needed to be modified to recognize isSponsorAddress functionality.

fixes rSmart/issues#201